contracts/requirements.txt
py-algorand-sdk>=2.5.0
algokit-utils>=2.0.0
python-dotenv>=1.0.0
